requirements.txt
django-admin-rangefilter==0.7.0
django-admin-numeric-filter==0.1.6
django-grappelli==2.14.3
django-admin-interface==0.14.2
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
six==1.15.0
defusedxml==0.6.0 
diff-match-patch==20200713
django-import-export==2.5.0
et-xmlfile==1.0.1
jdcal==1.4.1
markuppy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pyyaml==5.3.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
djadmin
suit

